Fix erts_bld_string_n (called by enif_make_string and erlang:decode_packet/3) to work with 8 bits characters

+++ erts/emulator/beam/utils.c
@@ -510,7 +510,7 @@ erts_bld_string_n(Uint **hpp, Uint *szp, const char *str, Sint len)
     if (hpp) {
 	res = NIL;
 	while (--i >= 0) {
-	    res = CONS(*hpp, make_small(str[i]), res);
+	    res = CONS(*hpp, make_small((byte) str[i]), res);
 	    *hpp += 2;
 	}
     }
